+++ studio/src/pages/[organizationSlug]/subgraphs.tsx
@@ -1,5 +1,4 @@
 import { useApplyParams } from "@/components/analytics/use-apply-params";
-import { NamespaceSelector } from "@/components/dashboard/NamespaceSelector";
 import { EmptyState } from "@/components/empty-state";
 import { getDashboardLayout } from "@/components/layout/dashboard-layout";
 import { SubgraphPageTabs, SubgraphsTable } from "@/components/subgraphs-table";
@@ -18,10 +17,12 @@ import {
 import { useRouter } from "next/router";
 import { useState } from "react";
 import { useDebounce } from "use-debounce";
+import { WorkspaceSelector } from "@/components/dashboard/workspace-selector";
+import { useWorkspace } from "@/hooks/use-workspace";
 
 const SubgraphsDashboardPage: NextPageWithLayout = () => {
   const router = useRouter();
-  const namespace = router.query.namespace as string;
+  const { namespace: { name: namespace } } = useWorkspace();
   const tab = router.query.tab as string;
 
   const pageNumber = router.query.page
@@ -37,7 +38,7 @@ const SubgraphsDashboardPage: NextPageWithLayout = () => {
   const applyParams = useApplyParams();
 
   const { data, isLoading, error, refetch } = useQuery(getSubgraphs, {
-    namespace: namespace || "default",
+    namespace: namespace,
     query,
     limit,
     offset,
@@ -158,7 +159,7 @@ SubgraphsDashboardPage.getLayout = (page) => {
     "An overview of all subgraphs",
     undefined,
     undefined,
-    [<NamespaceSelector key="0" />],
+    [<WorkspaceSelector key="0" />],
   );
 };