+++ shared/CHANGELOG.md
@@ -4,6 +4,18 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.42.4](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.42.3...@wundergraph/cosmo-shared@0.42.4) (2025-09-22)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
+## [0.42.3](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.42.2...@wundergraph/cosmo-shared@0.42.3) (2025-09-16)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
+## [0.42.2](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.42.1...@wundergraph/cosmo-shared@0.42.2) (2025-09-16)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
 ## [0.42.1](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.42.0...@wundergraph/cosmo-shared@0.42.1) (2025-09-10)
 
 **Note:** Version bump only for package @wundergraph/cosmo-shared